op-deployer: Fix intent validation with pre-deployed OPCM (#15712)
* op-deployer: Fix intent validation with pre-deployed OPCM

Previously, users were always required to provide the `SuperchainRoles` in their intent. Now that the intent supports providing an OPCM address, this is no longer necessary since we can look those values up on chain. The lookup was implemented in #15623, however users still had to provide `SuperchainRoles` in their intent to avoid validation errors.

The new logic is as follows:

- `standard` intents **must** specify the correct, standard OPCM address for the specified L1 and contracts version. Anything else is an error.
- All other intent types must **either** specify `SuperchainRoles`, or an OPCM address. Specifying both or neither will result in an error.
- The deployment pipeline will always pull `SuperchainRoles` data from the chain if the OPCM address is specified. Previously, it would only do this if the OPCM address was specified _and_ the L1 locator was a tag.

* cr updates