WIP
so far the interop-mon is starting up, and picking up information and logging metrics updates

I have not yet managed to wire up the interop_happy_tx_test.go so that it has a handle on the running interop mon .Part of the difficulty is handling the different backends. When it works I will be able to query the prometheus endpoint from in the test and make assertions about it.